+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -373,6 +373,8 @@ public class SampleSQLBuilder {
 		).put(
 			"messageBoardDataFactory", new MessageBoardDataFactory()
 		).put(
+			"releaseDataFactory", new ReleaseDataFactory()
+		).put(
 			"userDataFactory", new UserDataFactory()
 		).put(
 			"wikiDataFactory", new WikiDataFactory()
